refactor plugin folder structure
