+++ go/cmd/apiserver/config/base.yaml
@@ -1,7 +1,7 @@
 apiserver:
   yarpc:
     host: "localhost"
-    port: 14566
+    port: 14567
   crdSync:
     enableCRDUpdate: true
     enableIncompatibleUpdate: false
