+++ go/cmd/controllermgr/config/base.yaml
@@ -1,6 +1,6 @@
 controllermgr:
-  metricsBindAddress: ${METRICS_ADDRESS::8091}
-  healthProbeBindAddress: ${HEALTH_ADDRESS::8081}
+  metricsBindAddress: ${METRICS_ADDRESS::8083}
+  healthProbeBindAddress: ${HEALTH_ADDRESS::8084}
   leaderElection: false
   leaderElectionID: decaf1259.michelangelo.uber.com
   port: 9443
